mz668: add TestTarFileBuildcontext, factor shared tarball-context helpers
